Implement more advanced template for retros
Retrospectives now only appear next to the pair for the currently logged
in user. Add PageView methods to deal with this.

Model logic is in place to make the link on the pair page direct to
either the new or edit routes for pair_retros based on whether a retro
exists for that specific pair, user, year, and week.

Remove delete route for pair retros.

Make PairRetros delete automatically when the associated pair or user is
deleted (this includes repairifying, even if nothing changed)