Remove test for display of projects with no group
projects that do not have an associated group are no longer intended to
be displayed in the index.

Modify the add_group_to_project migration to allow groups to be deleted
without deleting associated projects so that users still have a record
of the projects theyve worked on even if the group was deleted in the
past.