Disallow creators changing project if not in group
Creators of projects cannot edit, update, or delete a project if they
are no longer in the group.

Projects can be changed by the owner of the group associated with the
project in all cases (even if they're not in the users association of
the group).

Projects cannot be changed the creator of a project if they are no
longer in the project's associated group.

Projects can be changed by the creator of a project if they are still in
the associated group.

Otherwise, projects cannot be changed. This includes other members of
the associated group that are not either the creator of the project or
the owner of the associated group.