Fix updating retros with new users or pairs
Fixes manually being able to hit the :update route to update the user or
pair of a retrospective by removing that from the cast parameters
available in an update action to the changeset.

Fix being able to update the pair_date to a date before the date of a
pair if the pair_id parameter was not explicitly supplied.

Add pair and retro ex machina factories

Refactor pair_retro controller tests to use more ex machina and divide
out the actions into their own describe blocks