Add project associated with retro validation
Make the retro changeset validate that the passed in project is
associated with the group that the pair that the retro is associated
with is associated with (what a sentence)

Make the pair_retro new form only list projects that the pair is
associated with through that pair's group.

Make the pair_retro :create action validate that the passed in project
is associated with the pair through the group.