Refactor PairController
Change the name of fetch_pairs to fetch_groups_and_pairs

Instead of passing the conn deep into fetch_groups_and_pairs and
pairs_for_user_groups, pass up the messages so that they can be attached
to the conn at the shallowest function in the controller.

Unpipe a few reduce calls to reduce indentation and make the code a bit
clearer.

Eliminate some variable declarations that were necessary because we were
passing the conn down.